feat: implementar acao de deletar review do usuario

+++ lib/src/domain/use_case/delete_review_use_case.dart
@@ -11,9 +11,12 @@ class DeleteReviewUseCase extends SafeUseCase {
     _service = Modular.get<LocationService>();
   }
 
-  Future<String> call(int idReview) async {
-    final response = await _service.deleteReview(idReview);
-
-    return response.message ?? S.current.textDefaultDeleteReviewMessage;
+  Future<String> call(int? idReview) async {
+    if (idReview != null) {
+      final response = await _service.deleteReview(idReview);
+      return response.message ?? S.current.textDefaultDeleteReviewMessage;
+    } else {
+      return 'Erro ao deletar a review, tente novamente em alguns instantes!';
+    }
   }
 }
